pyproject.toml
[tool.poetry]
name = "cdk-constructs"
version = "0.2.0"
description = "Validate pull requests"
authors = ["Myself"]

[tool.poetry.dependencies]
python = "^3.7"
"aws-cdk.aws-lambda" = "^1.73.0"
"aws-cdk.aws-codepipeline" = "^1.73.0"
"aws-cdk.aws-codebuild" = "^1.73.0"
"aws-cdk.aws-codepipeline-actions" = "^1.73.0"
"aws-cdk.aws-events" = "^1.73.0"
"aws-cdk.aws-events-targets" = "^1.73.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
